requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
pyaes==1.6.1
pyasn1==0.6.1
Pyrogram==2.0.106
PySocks==1.7.1
requests==2.32.5
rsa==4.9.1
Telethon==1.40.0
urllib3==2.5.0